(_metric_to_line): convert all timestamps to long integers before transmitting to Carbon.
The regular carbon-cache/carbon-relay daemons do accept floating point
timestamps, but the drop-in replacement carbon-cache-ng will refuse to accept
any metric with a non-integer timestamp. Who needs the extra precision, right?

+++ python/nav/metrics/carbon.py
@@ -129,7 +129,7 @@ def _socktype_from_addr(addr):
 
 def _metric_to_line(metric_tuple):
     path, (timestamp, value) = metric_tuple
-    return str("%s %s %s\n" % (path, value, timestamp))
+    return str("%s %s %s\n" % (path, value, long(timestamp)))
 
 
 def metrics_to_packets(metric_tuples):
